Revert "Pressing enter will insert new line at current cursor position"
This reverts commit d010920ea5932cf5291c5d79b843ae8ee2ee4def.

This causes problems when the cursor is at the end of line, and we just
want to insert a new line. Instead of doing that, it takes the last
character of the line with it, because in normal mode the cursor can't
go past the last character.

Let's just keep it simple.