Use vim-commentary instead of my own commenting mechanism
It is much better. Firstly, it is a toggle so I don't have to think
which bind to use. It is the same for everything. Plus, it seems to work
out of the box for most file types I've tried, so I don't have to
manually add too many comment types in my vimrc.